+++ cmd/status.go
@@ -1,14 +1,11 @@
 package cmd
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"strings"
 	"text/tabwriter"
 
-	"github.com/manifoldco/torus-cli/api"
-	"github.com/manifoldco/torus-cli/config"
 	"github.com/manifoldco/torus-cli/errs"
 	"github.com/manifoldco/torus-cli/prefs"
 	"github.com/manifoldco/torus-cli/ui"
@@ -62,19 +59,6 @@ func statusCmd(ctx *cli.Context) error {
 		return errs.NewExitError(msg)
 	}
 
-	cfg, err := config.LoadConfig()
-	if err != nil {
-		return err
-	}
-
-	client := api.NewClient(cfg)
-	c := context.Background()
-
-	session, err := client.Session.Who(c)
-	if err != nil {
-		return errs.NewErrorExitError("Error fetching identity", err)
-	}
-
 	err = checkRequiredFlags(ctx)
 	if err != nil {
 		fmt.Printf("You are not inside a linked working directory. "+
@@ -82,27 +66,19 @@ func statusCmd(ctx *cli.Context) error {
 		return nil
 	}
 
-	identity, err := deriveIdentity(ctx, session)
-	if err != nil {
-		return err
-	}
-
 	org := ctx.String("org")
 	project := ctx.String("project")
 	env := ctx.String("environment")
 	service := ctx.String("service")
-	instance := ctx.String("instance")
 
 	w := tabwriter.NewWriter(os.Stdout, 2, 0, 1, ' ', 0)
 	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Org"), org)
 	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Project"), project)
 	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Environment"), env)
 	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Service"), service)
-	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Identity"), ui.FaintString(identity))
-	fmt.Fprintf(w, "%s:\t%s\n", ui.BoldString("Instance"), instance)
 	w.Flush()
 
-	parts := []string{"", org, project, env, service, identity, instance}
+	parts := []string{"", org, project, env, service}
 	credPath := strings.Join(parts, "/")
 	fmt.Printf("\n%s: %s\n", ui.BoldString("Credential Path"), credPath)